tools/domain-risk-scan/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
alembic
celery
redis
dnspython
httpx
python-dotenv
jinja2
stripe
reportlab